services/gateway/requirements.txt
# Django Libs
Django==3.2.5
djangorestframework==3.12.4
django-cors-headers==3.5.0
psycopg2==2.9.1
requests==2.26.0

# Kafka Libs
faust[rocksdb]==1.10.4
eventlet==0.31.1
kafka-python==1.4.7